Supporting Python from 2.6 to 3.3.
This commit adds python 2.6-3.3 support (as well as tests through tox) to slumber.

Some additional dependencies were introduced, though. Six was added to support conditional treatment of strings.

A test dependency to unittest2 was added to tox in order to add support to the "discover" method of TestCase (though this dependency only happens in python26).

All tests are passing in the following python versions: 2.6, 2.7, 3.2, 3.3 and pypy.